Add PostCSS and Tailwind CSS configuration and dependencies
This commit adds PostCSS and Tailwind CSS to the project. Two new configuration files for PostCSS and Tailwind have been created and relevant dependencies have been added to the package.json file. This change introduces the use of Tailwind for styling the application and PostCSS for handling CSS transformations. Various changes in other files result from the application of these new tools.