Update layout and navigation module with routing functionality
This commit introduces improved routing functionality by adding 'routeLoader$' and 'useNavigate' to the layout module along with the 'RouteNavigate' type to the navigation component. The layout now dynamically retrieves the page pathname and passes it along with the navigation function to the Navigation component via props. Furthermore, an instance of 'NextUIProvider' now encapsulates the 'Navbar' component in the navigation module, allowing enhanced link navigation throughout the application.