Add new routes and update page loading from Sanity queries
Implemented new routes for generic pages, page gallery, and pages overview in the application. These routes contain components to fetch and display data. A redirect for the admin page was added in '_redirects'. Updated the 'getPage' queries function in 'get-page.ts' from Sanity to remove the NO_DRAFTS filter for loading pages.